+++ CHANGELOG.md
@@ -5,6 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.14.2] - 2023-07-24
+
+### Fixed
+* DAG mapping incorrectly handling stages with missing inputs in state.
+
+
+## [0.14.1] - 2023-07-21
+
+### Fixed
+* DAG never adding a stage with no outputs to the execution list.
+
+
+
+
 ## [0.14.0] - 2023-07-19
 
 DAG-based execution of stages is finally here!